+++ src/orquestra/sdk/secrets/_api.py
@@ -18,6 +18,14 @@ def _translate_to_zri(workspace_id: str, secret_name: str) -> str:
     return f"zri:v1::0:{workspace_id}:secret:{secret_name}"
 
 
+def _raise_warning_for_none_workspace(workspace: t.Optional[str]):
+    if workspace is None:
+        raise FutureWarning(
+            "Please specify workspace ID directly for accessing secrets."
+            " Support for default workspaces will be sunset in the future."
+        )
+
+
 def get(
     name: str,
     *,
@@ -50,6 +58,8 @@ def get(
             this function will return a "future" which will be used to retrieve the
             secret at execution time.
     """
+    _raise_warning_for_none_workspace(workspace_id)
+
     if _exec_ctx.global_context == _exec_ctx.ExecContext.WORKFLOW_BUILD:
         return t.cast(
             str,
@@ -96,6 +106,8 @@ def list(
         orquestra.sdk.exceptions.UnauthorizedError: when the authorization with the
             remote vault failed.
     """
+    _raise_warning_for_none_workspace(workspace_id)
+
     try:
         client = _auth.authorized_client(config_name)
     except sdk_exc.ConfigNameNotFoundError:
@@ -133,6 +145,8 @@ def set(
         orquestra.sdk.exceptions.UnauthorizedError: when the authorization with the
             remote vault failed.
     """
+    _raise_warning_for_none_workspace(workspace_id)
+
     try:
         client = _auth.authorized_client(config_name)
     except sdk_exc.ConfigNameNotFoundError:
@@ -179,6 +193,8 @@ def delete(
         orquestra.sdk.exceptions.UnauthorizedError: when the authorization with the
             remote vault failed.
     """
+    _raise_warning_for_none_workspace(workspace_id)
+
     try:
         client = _auth.authorized_client(config_name)
     except sdk_exc.ConfigNameNotFoundError: